requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
pandas==2.2.0
pyreadstat==1.2.7
aiohttp==3.10.0
google-generativeai==0.8.0
python-multipart==0.0.9
sse-starlette==2.1.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
slowapi==0.1.9
stripe==10.0.0
openpyxl==3.1.5
reportlab==4.2.0
yfinance==0.2.40
